Prim's MST: Actually use a priority queue.
In my defence, the code was old, from when I first learned the
algorithm and had just learned what it does, not how it is usually done
with a priority queue (binary heap). Note that it should really be done
with a Fibonnaci heap so we can use its O(1) time decrease-key operation.